move: config 파일 분리(이동)
- 이후 config 파일을 더 추가하는 것을 대비해 config 파일은 따로 분리함
- 토큰 관련 파일은 의미를 명확히 하기 위해 'token'으로 패키지명 변경
