+++ server/src/main/java/com/codecozy/server/controller/HomeController.java
@@ -1,6 +1,6 @@
 package com.codecozy.server.controller;
 
-import com.codecozy.server.security.TokenProvider;
+import com.codecozy.server.token.TokenProvider;
 import com.codecozy.server.service.HomeService;
 import java.io.IOException;
 import java.util.Map;
