+++ server/src/main/java/com/codecozy/server/token/JwtAuthenticateFilter.java
@@ -1,4 +1,4 @@
-package com.codecozy.server.security;
+package com.codecozy.server.token;
 
 import com.codecozy.server.context.ResponseMessages;
 import com.codecozy.server.context.StatusCode;
